Переименованы компоненты страниц
